# youthchinaopenapi.yml
openapi: 3.0.0
info:
  version: 1.5.3
  title: YouthChina
  license:
    name: MIT
servers:
  - url: 'http://youthchina.io/api/v1'
paths:
...
  '/jobs/{id}/apply/resume':
    post:
      tags:
      - Application
      description: send resume to company
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: string
              format: binary
              description: video to uploaded
      parameters:
        - name: id
          in: path
          description: id of the job
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: successfully send email
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    $ref: '#/components/schemas/Application'
                  status:
                    $ref: '#/components/schemas/Status'
...
  '/videos/{id}':
    get:
      tags:
        - Video
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    $ref: '#/components/schemas/Video'
                  status:
                    $ref: '#/components/schemas/Status'
    delete:
      tags:
        - Video
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true  
      responses:
        '204':
          description: successful
...
  '/videos/{id}/comments':
    post:
      tags:
        - Video
      parameters:
        - name: id
          schema:
            type: integer
          in: path
          required: true
          description: video id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
      responses:
        '201':
          description: successful
    get:
      tags:
        - Video
      parameters:
        - name: id
          schema:
            type: integer
          in: path
          required: true
          description: video id
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: object
                    properties:
                      comments:
                        type: array
                        items:
                          $ref: '#/components/schemas/Comment'
                  status:
                    $ref: '#/components/schemas/Status'
  '/videos/{id}/upvote':
    put:  
      tags:
        - Video
      parameters:
        - name: id
          schema:
            type: integer
          in: path
          required: true
          description: video id
      responses:
        '200':
          description: successful 
  '/videos/{id}/attention':
    put:
      tags:
        - Video
      parameters:
        - name: id
          schema:
            type: integer
          in: path
          required: true
          description: video id
      responses:
        '200':
          description: successful 
...
  '/resumes':
    post:
      tags:
        - Resume
      requestBody:
       content:
         application/json:
             schema:
               type: object
               properties:
                 resume:
                   type: string
      responses:
        '201':
          description: created
          content:
           application/json:
              schema:
                type: object
                properties:
                  content:
                    type: object
                    properties:
                      id:
                        type: integer
                      jsons:
                        type: array
                        items:
                          type: object
                  status:
                    $ref: '#/components/schemas/Status'
  '/resumes/{id}':
    get:
      tags:
        - Resume
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: resume id
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: object
                    properties:
                      id:
                        type: integer
                      jsons:
                        type: array
                        items:
                          type: object
                  status:
                    $ref: '#/components/schemas/Status'
    delete:
      tags:
      - Resume
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '204':
          description: deleted
  '/applicant/{id}/resumes':
    get:
      tags:
        - applicant
      parameters:
        - name: id
          in: path
          description: applicant id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: resumes
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: object
                    properties:
                      resumes:
                        type: object
                        properties:
                          id:
                            type: integer
                          jsons:
                            type: array
                            items:
                              type: object
                  status:
                    $ref: '#/components/schemas/Status'
...
components:
  schemas:
    ContentRequest:
      type: object
      properties:
        body:
          $ref: '#/components/schemas/RichText'
    Status:
      type: object
      properties:
        code:
          type: integer
          format: int64
        reason:
          type: string
...
    Job:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          description: name of job
        organization:
          $ref: '#/components/schemas/Organization'
        location:
          type: string
        type:
          type: string
          enum:
            - full-time
            - part-time
            - internship
        deadLine:
          type: string
        job_description:
          type: string
        job_duty:
          type: string
    Organization:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        avatarUrl:
          type: string
          pattern: '(http|https)://(.?)*'
        location:
          type: string
        website:
          type: string
          pattern: '(http|https)://(.?)*'
        note:
          type: string
        nation:
          type: string
    User:
      type: object
      required:
        - id
        - username
        - email
        - phonenumber
        - register_date
        - real_name
        - gender
        - nation
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        phonenumber:
          type: string
        register_date:
          type: string
        real_name:
          type: string
        gender:
          type: string
          enum:
            - male
            - female
        nation:
          type: string
        avatar_url:
          type: string
        role:
          type: string
        age:
          type: integer
...
    Application:
      type: object
      required:
        - id
        - companyId
        - positionId
        - status
      properties:
        id:
          type: integer
        position:
          $ref: '#/components/schemas/Job'
        status:
          type: string
...
    RichText:
      type: object
      required:
        - braftEditorRaw
        - previewText
        - resourceList
      properties:
        braftEditorRaw:
          type: object
        previewText:
          type: object
        resourceList:
          type: array
          items: 
            type: string
    Comment:
      type: object
      required:
        - id
        - creator
        - body
        - create_at
        - is_anonymous
      properties:
        id:
          type: integer
        creator:
          $ref: '#/components/schemas/User'
        body:
          $ref: '#/components/schemas/RichText'
        create_at:
          type: integer
        is_anonymous:
          type: boolean
          default: false
    CommentRequest:
      allOf:
        - $ref: '#/components/schemas/ContentRequest'
        - properties:
            is_anonymous:
              type: boolean
              default: false
...
    Video:
      type: object
      properties:
        id:
          type: integer
          description: video id
        url:
          type: string
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        uploader:
          $ref: '#/components/schemas/User'
...